requirements.txt
# VideoSDK AI Agents Core
videosdk-agents==0.0.24

# Real-time API and LLM-based TTS/STT Providers
videosdk-plugins-openai>=0.0.24
videosdk-plugins-google>=0.0.24
videosdk-plugins-aws>=0.0.24
videosdk-plugins-sarvamai>=0.0.24


# Speech-to-Text (STT) Plugins
videosdk-plugins-deepgram>=0.0.24


# Voice Activity Detection (VAD) and Turn Detection Plugins
videosdk-plugins-silero>=0.0.24
videosdk-plugins-turn-detector>=0.0.24

# Text-to-Speech (TTS) Plugins
videosdk-plugins-elevenlabs>=0.0.24

# LLM Plugins
videosdk-plugins-anthropic>=0.0.24

# Virtual Avatar Plugins
videosdk-plugins-simli>=0.0.24

# Core Dependencies
fastmcp
aiohttp
